beanlibapi/apps/core/permissions.py
from rest_framework import permissions


class IsOwnerOrReadOnly(permissions.BasePermission):
    def has_permission(self, request, view):
        return request.user and request.user.is_authenticated

    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True

        print(f"obj.owner: {obj.owner}")
        print(f"request.user: {request.user}")

        return obj.owner == request.user
