pyproject.toml
[tool.poetry]
name = "bean-library-api"
version = "0.1.0"
description = ""
authors = ["ask4git <ask4git@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.16"
boto3 = "^1.35.39"
mysqlclient = "^2.2.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
dj-rest-auth = "^6.0.0"
requests = "^2.32.3"
shortid = "^0.1.2"
gunicorn = "^23.0.0"
django-cors-headers = "^4.5.0"
django-bootstrap4 = "^24.4"
django-storages = "^1.14.4"
django-environ = "^0.11.2"
dj-database-url = "^2.2.0"
psycopg = "^3.2.3"
pytest = "^8.3.3"
cryptography = "^43.0.3"
drf-spectacular = "^0.27.2"
drf-rw-serializers = "^1.4.0"
django-allauth = "0.61.1"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
console_output_style = "progress"
testpaths = [
    "beanlibapi/apps/authx/tests",
    "beanlibapi/apps/core/tests",
]